refactor: context should not be coupled in controller
